src/energy_cross_border_flows/db_manager/requirements.txt
pandas==1.1.5
SQLAlchemy==1.3.23
fastapi==0.66.0
uvicorn==0.14.0
numpy==1.19.5
psycopg2-binary==2.8.6
PyYAML==5.4.1       
